Stop using URLValidator to validate return URLs.
It's excessively picky, and in particular rejects single-component host
names.  These are permitted by the URI spec, and are generated by
django.test.RequestFactory.  In any case, the next thing we do with the
returned URL is to compare it with one we've generated internally,
which should catch any genuinely invalid URLs.